+++ src/components/base/Base.js
@@ -1,12 +1,13 @@
-import maskInput from 'vanilla-text-mask';
+import maskInput, { conformToMask } from 'vanilla-text-mask';
 import Promise from "native-promise-only";
-import _ from 'lodash';
+import _set from 'lodash/set';
 import _get from 'lodash/get';
 import _each from 'lodash/each';
 import _assign from 'lodash/assign';
 import _debounce from 'lodash/debounce';
 import _isArray from 'lodash/isArray';
 import _clone from 'lodash/clone';
+import _cloneDeep from 'lodash/cloneDeep';
 import _defaults from 'lodash/defaults';
 import _isEqual from 'lodash/isEqual';
 import _isUndefined from 'lodash/isUndefined';
@@ -28,6 +29,7 @@ export class BaseComponent {
    * @param {Object} data - The global data submission object this component will belong.
    */
   constructor(component, options, data) {
+    this.originalComponent = _cloneDeep(component);
     /**
      * The ID of this component. This value is auto-generated when the component is created, but
      * can also be provided from the component.id value passed into the constructor.
@@ -170,7 +172,7 @@ export class BaseComponent {
      * The validators that are assigned to this component.
      * @type {[string]}
      */
-    this.validators = ['required', 'minLength', 'maxLength', 'custom', 'pattern', 'json'];
+    this.validators = ['required', 'minLength', 'maxLength', 'custom', 'pattern', 'json', 'mask'];
 
     /**
      * Used to trigger a new change in this component.
@@ -186,6 +188,9 @@ export class BaseComponent {
 
     // To force this component to be invalid.
     this.invalid = false;
+
+    // Determine if the component has been built.
+    this.isBuilt = false;
 
     /**
      * An array of the event listeners so that the destroy command can deregister them.
@@ -482,16 +487,19 @@ export class BaseComponent {
    * @returns {HTMLElement}
    */
   createElement() {
+    // If the element is already created, don't recreate.
+    if (this.element) {
+      return this.element;
+    }
+
     this.element = this.ce('div', {
       id: this.id,
       class: this.className,
       style: this.customStyle
     });
 
-    if (this.element) {
-      // Ensure you can get the component info from the element.
-      this.element.component = this.component;
-    }
+    // Ensure you can get the component info from the element.
+    this.element.component = this.component;
 
     return this.element;
   }
@@ -535,12 +543,7 @@ export class BaseComponent {
     else if (this.component.customDefaultValue) {
       if (typeof this.component.customDefaultValue === 'string') {
         try {
-          let row = this.data;
-          let data = this.data;
-          let value = '';
-          let component = this;
-          eval(this.component.customDefaultValue.toString());
-          defaultValue = value;
+          defaultValue = (new Function('component', 'row', 'data', `var value = ''; ${this.component.customDefaultValue.toString()}; return value;`))(this, this.data, this.data);
         }
         catch (e) {
           defaultValue = null;
@@ -553,8 +556,7 @@ export class BaseComponent {
         try {
           defaultValue = FormioUtils.jsonLogic.apply(this.component.customDefaultValue, {
             data: this.data,
-            row: this.data,
-            _
+            row: this.data
           });
         }
         catch (err) {
@@ -565,6 +567,14 @@ export class BaseComponent {
         }
       }
     }
+
+    if (this._inputMask) {
+      defaultValue = conformToMask(defaultValue, this._inputMask).conformedValue;
+      if (!FormioUtils.matchInputMask(defaultValue, this._inputMask)) {
+        defaultValue = '';
+      }
+    }
+
     return defaultValue;
   }
 
@@ -1034,50 +1044,12 @@ export class BaseComponent {
   }
 
   /**
-   * Returns an input mask that is compatible with the input mask library.
-   * @param {string} mask - The Form.io input mask.
-   * @returns {Array} - The input mask for the mask library.
-   */
-  getInputMask(mask) {
-    if (mask instanceof Array) {
-      return mask;
-    }
-    let maskArray = [];
-    maskArray.numeric = true;
-    for (let i=0; i < mask.length; i++) {
-      switch (mask[i]) {
-        case '9':
-          maskArray.push(/\d/);
-          break;
-        case 'A':
-          maskArray.numeric = false;
-          maskArray.push(/[a-zA-Z]/);
-          break;
-        case 'a':
-          maskArray.numeric = false;
-          maskArray.push(/[a-z]/);
-          break;
-        case '*':
-          maskArray.numeric = false;
-          maskArray.push(/[a-zA-Z0-9]/);
-          break;
-        default:
-          maskArray.push(mask[i]);
-          break;
-      }
-    }
-    return maskArray;
-  }
-
-  /**
    * Creates a new input mask placeholder.
    * @param {HTMLElement} mask - The input mask.
    * @returns {string} - The placeholder that will exist within the input as they type.
    */
   maskPlaceholder(mask) {
-    return mask.map((char) => {
-      return (char instanceof RegExp) ? '_' : char
-    }).join('')
+    return mask.map((char) => (char instanceof RegExp) ? '_' : char).join('');
   }
 
   /**
@@ -1086,10 +1058,11 @@ export class BaseComponent {
    */
   setInputMask(input) {
     if (input && this.component.inputMask) {
-      let mask = this.getInputMask(this.component.inputMask);
+      const mask = FormioUtils.getInputMask(this.component.inputMask);
+      this._inputMask = mask;
       this.inputMask = maskInput({
         inputElement: input,
-        mask: mask
+        mask
       });
       if (mask.numeric) {
         input.setAttribute('pattern', "\\d*");
@@ -1138,6 +1111,10 @@ export class BaseComponent {
   }
 
   redraw() {
+    // Don't bother if we have not built yet.
+    if (!this.isBuilt) {
+      return;
+    }
     this.clear();
     this.build();
   }
@@ -1159,6 +1136,7 @@ export class BaseComponent {
         window.removeEventListener(handler.event, handler.func);
       }
     });
+    this.inputs = [];
   }
 
   /**
@@ -1269,7 +1247,10 @@ export class BaseComponent {
    *   The name of the class to add.
    */
   addClass(element, className) {
-    element.setAttribute('class', `${element.getAttribute('class')} ${className}`);
+    const classes = element.getAttribute('class');
+    if (!classes || classes.indexOf(className) === -1) {
+      element.setAttribute('class', `${classes} ${className}`);
+    }
   }
 
   /**
@@ -1283,7 +1264,7 @@ export class BaseComponent {
   removeClass(element, className) {
     let cls = element.getAttribute('class');
     if (cls) {
-      cls = cls.replace(className, '');
+      cls = cls.replace(new RegExp(className, 'g'), '');
       element.setAttribute('class', cls);
     }
   }
@@ -1306,11 +1287,71 @@ export class BaseComponent {
    * Check for conditionals and hide/show the element based on those conditions.
    */
   checkConditions(data) {
+    // Check advanced conditions
+    let result;
+
     if (!this.hasCondition()) {
-      return this.show(true);
+      result = this.show(true);
+    }
+    else {
+      result = this.show(FormioUtils.checkCondition(this.component, this.data, data));
     }
 
-    return this.show(FormioUtils.checkCondition(this.component, this.data, data));
+    if (this.fieldLogic(data)) {
+      this.redraw();
+    }
+
+    return result;
+  }
+
+  /**
+   * Check all triggers and apply necessary actions.
+   *
+   * @param data
+   */
+  fieldLogic(data) {
+    const logics = this.component.logic || [];
+
+    // If there aren't logic, don't go further.
+    if (logics.length === 0) {
+      return;
+    }
+
+    const newComponent = _cloneDeep(this.originalComponent);
+
+    let changed = logics.reduce((changed, logic) => {
+      const result = FormioUtils.checkTrigger(newComponent, logic.trigger, this.data, data);
+
+      if (result) {
+        changed |= logic.actions.reduce((changed, action) => {
+          switch(action.type) {
+            case 'property':
+              FormioUtils.setActionProperty(newComponent, action, this.data, data, newComponent, result);
+              break;
+            case 'value':
+              const newValue = (new Function('row', 'data', 'component', 'result', action.value))(this.data, data, newComponent, result);
+              if (!_isEqual(this.getValue(), newValue)) {
+                this.setValue(newValue);
+                changed = true;
+              }
+              break;
+            case 'validation':
+              // TODO
+              break;
+          }
+          return changed;
+        }, false);
+      }
+      return changed;
+    }, false);
+
+    // If component definition changed, replace and mark as changed.
+    if (!_isEqual(this.component, newComponent)) {
+      this.component = newComponent;
+      changed = true;
+    }
+
+    return changed;
   }
 
   /**
@@ -1580,15 +1621,12 @@ export class BaseComponent {
     // If this is a string, then use eval to evalulate it.
     if (typeof this.component.calculateValue === 'string') {
       try {
-        let value = [];
-        let row = this.data;
-        let component = this;
-        eval(this.component.calculateValue.toString());
+        let value = (new Function('component', 'row', `value = []; ${this.component.calculateValue.toString()}; return value;`))(this, this.data);
         changed = this.setValue(value, flags);
       }
-      catch (e) {
+      catch (err) {
         /* eslint-disable no-console */
-        console.warn(`An error occurred calculating a value for ${this.component.key}`, e);
+        console.warn(`An error occurred calculating a value for ${this.component.key}`, err);
         changed = false;
         /* eslint-enable no-console */
       }
@@ -1597,14 +1635,13 @@ export class BaseComponent {
       try {
         let val = FormioUtils.jsonLogic.apply(this.component.calculateValue, {
           data,
-          row: this.data,
-          _
+          row: this.data
         });
         changed = this.setValue(val, flags);
       }
       catch (err) {
         /* eslint-disable no-console */
-        console.warn(`An error occurred calculating a value for ${this.component.key}`, e);
+        console.warn(`An error occurred calculating a value for ${this.component.key}`, err);
         changed = false;
         /* eslint-enable no-console */
       }
@@ -1887,7 +1924,9 @@ export class BaseComponent {
     this.destroy();
     const element = this.getElement();
     if (element) {
-      element.innerHTML = '';
+      while (element.lastChild) {
+        element.removeChild(element.lastChild);
+      }
     }
   }
 
@@ -1923,7 +1962,7 @@ export class BaseComponent {
       type: this.component.inputType || 'text',
       class: 'form-control',
       lang: this.options.language
-  };
+    };
 
     if (this.component.placeholder) {
       attributes.placeholder = this.t(this.component.placeholder);
@@ -1931,6 +1970,10 @@ export class BaseComponent {
 
     if (this.component.tabindex) {
       attributes.tabindex = this.component.tabindex;
+    }
+
+    if (this.component.autofocus) {
+      attributes.autofocus = this.component.autofocus;
     }
 
     return {