+++ src/components/form/Form.js
@@ -1,33 +1,39 @@
 import FormioForm from '../../formio.form';
 import FormioUtils from '../../utils';
 import Formio from '../../formio';
-import _merge from 'lodash/merge';
 import _isEmpty from 'lodash/isEmpty';
 
 export class FormComponent extends FormioForm {
   constructor(component, options, data) {
     data = data || {};
-
-    // We need to reset the language so that it will not try to rebuild the form (by setting the language)
-    // before this form component is done rendering. We will set it later.
-    let language = '';
-    if (options && options.language) {
-      language = options.language;
-      options.language = '';
-    }
     super(null, options);
-    if (language && options) {
-      options.language = language;
-    }
 
     // Ensure this component does not make it to the global forms array.
     delete Formio.forms[this.id];
-
     this.type = 'formcomponent';
     this.component = component;
     this.submitted = false;
     this.data = data;
+    this.readyPromise = new Promise((resolve) => {
+      this.readyResolve = resolve;
+    });
+  }
+
+  /**
+   * Load the subform.
+   */
+  loadSubForm() {
+    if (this.subFormLoaded) {
+      return true;
+    }
+    this.subFormLoaded = true;
     let srcOptions = {};
+    if (this.options && this.options.base) {
+      srcOptions.base = this.options.base;
+    }
+    if (this.options && this.options.project) {
+      srcOptions.project = this.options.project;
+    }
 
     // Make sure that if reference is provided, the form must submit.
     if (this.component.reference) {
@@ -35,76 +41,88 @@ export class FormComponent extends FormioForm {
     }
 
     if (
-      !component.src &&
+      !this.component.src &&
       !this.options.formio &&
-      component.form
+      this.component.form
     ) {
-      component.src = Formio.getBaseUrl();
-      if (component.project) {
+      this.component.src = Formio.getBaseUrl();
+      if (this.component.project) {
         // Check to see if it is a MongoID.
-        if (FormioUtils.isMongoId(component.project)) {
-          component.src += '/project';
+        if (FormioUtils.isMongoId(this.component.project)) {
+          this.component.src += '/project';
         }
-        component.src += '/' + component.project;
-        srcOptions.project = component.src;
+        this.component.src += '/' + this.component.project;
+        srcOptions.project = this.component.src;
       }
-      component.src += '/form/' + component.form;
+      this.component.src += '/form/' + this.component.form;
     }
 
     // Build the source based on the root src path.
-    if (!component.src && this.options.formio) {
+    if (!this.component.src && this.options.formio) {
       let rootSrc = this.options.formio.formsUrl;
-      if (component.path) {
+      if (this.component.path) {
         let parts = rootSrc.split('/');
         parts.pop();
-        component.src = parts.join('/') + '/' + component.path;
+        this.component.src = parts.join('/') + '/' + this.component.path;
       }
-      if (component.form) {
-        component.src = rootSrc + '/' + component.form;
+      if (this.component.form) {
+        this.component.src = rootSrc + '/' + this.component.form;
       }
     }
 
     // Add the source to this actual submission if the component is a reference.
     if (
-      data &&
-      data[component.key] &&
-      data[component.key]._id &&
+      this.data &&
+      this.data[this.component.key] &&
+      this.data[this.component.key]._id &&
       this.component.reference &&
-      !component.src.includes('/submission/')
+      !this.component.src.includes('/submission/')
     ) {
-      component.src += '/submission/' + data[component.key]._id;
+      this.component.src += '/submission/' + this.data[this.component.key]._id;
     }
 
     // Set the src if the property is provided in the JSON.
-    if (component.src) {
-      this.setSrc(component.src, srcOptions);
+    if (this.component.src) {
+      this.setSrc(this.component.src, srcOptions);
     }
 
     // Directly set the submission if it isn't a reference.
-    if (data && data[component.key] && !this.component.reference) {
-      this.setSubmission(data[component.key]);
+    if (this.data && this.data[this.component.key] && !this.component.reference) {
+      this.setSubmission(this.data[this.component.key]);
     }
-
-    this.readyPromise = new Promise((resolve) => {
-      this.readyResolve = resolve;
-    });
 
     // Set language after everything is established.
-    if (options && options.language) {
-      this.language = options.language;
+    if (this.options && this.options.language) {
+      this.language = this.options.language;
     }
   }
 
-  checkValidity(data) {
-    return super.checkValidity(data[this.component.key] ? data[this.component.key].data : {});
+  get subData() {
+    if (!this.data[this.component.key]) {
+      this.data[this.component.key] = {data: {}};
+    }
+    return this.data[this.component.key].data;
   }
 
-  checkConditions(data) {
-    return super.checkConditions(data[this.component.key] ? data[this.component.key].data : {});
+  checkValidity() {
+    // Maintain isolated data scope when passing root data for validity checks.
+    return super.checkValidity(this.subData);
+  }
+
+  checkConditions() {
+    // Check the conditions for the subform.
+    if (super.checkConditions(this.subData)) {
+      // Only load the subform if this component is visible.
+      this.loadSubForm();
+      return true;
+    }
+
+    return false;
   }
 
   calculateValue(data, flags) {
-    return super.calculateValue(data[this.component.key] ? data[this.component.key].data : {}, flags);
+    // Maintain isolated data scope when calculating values.
+    return super.calculateValue(this.subData, flags);
   }
 
   /**
@@ -206,6 +224,11 @@ export class FormComponent extends FormioForm {
       this.data[this.component.key] = this._submission = {data: {}};
       this.readyResolve();
       return;
+    }
+
+    // Load the subform if we have data.
+    if (submission._id || !_isEmpty(this.data[this.component.key])) {
+      this.loadSubForm();
     }
 
     // Set the url of this form to the url for a submission if it exists.