+++ src/components/textfield/TextField.js
@@ -4,6 +4,10 @@ export class TextFieldComponent extends BaseComponent {
     let info = super.elementInfo();
     info.type = 'input';
 
+    if (this.component.hasOwnProperty('spellcheck')) {
+      info.attr.spellcheck = this.component.spellcheck;
+    }
+
     if (this.component.mask) {
       info.attr.type = 'password';
     }