+++ src/formio.form.js
@@ -420,17 +420,18 @@ export class FormioForm extends FormioComponents {
    * @param options
    */
   setSrc(value, options) {
-    this.setUrl(value, options);
-    this.nosubmit = false;
-    this.formio.loadForm({params: {live: 1}}).then(
-      (form) => {
-        var setForm = this.setForm(form);
-        this.loadSubmission();
-        return setForm;
-      }).catch((err) => {
+    if (this.setUrl(value, options)) {
+      this.nosubmit = false;
+      this.formio.loadForm({params: {live: 1}}).then(
+        (form) => {
+          var setForm = this.setForm(form);
+          this.loadSubmission();
+          return setForm;
+        }).catch((err) => {
         console.warn(err);
         this.formReadyReject(err);
       });
+    }
   }
 
   /**
@@ -465,8 +466,12 @@ export class FormioForm extends FormioComponents {
    * @param options
    */
   setUrl(value, options) {
-    if (!value || typeof value !== 'string') {
-      return;
+    if (
+      !value ||
+      (typeof value !== 'string') ||
+      (value === this._src)
+    ) {
+      return false;
     }
     this._src = value;
     this.nosubmit = true;
@@ -476,6 +481,7 @@ export class FormioForm extends FormioComponents {
       // Set the options source so this can be passed to other components.
       this.options.src = value;
     }
+    return true;
   }
 
   /**
@@ -727,6 +733,7 @@ export class FormioForm extends FormioComponents {
       this.clear();
       return this.localize().then(() => {
         this.build();
+        this.isBuilt = true;
         this.onResize();
         this.on('resetForm', () => this.reset(), true);
         this.on('refreshData', () => this.updateValue());
@@ -779,6 +786,7 @@ export class FormioForm extends FormioComponents {
     this.checkData(submission.data, {
       noValidate: true
     });
+    this.on('requestUrl', (args) => (this.submitUrl(args.url,args.headers)), true);
   }
 
   /**
@@ -973,6 +981,45 @@ export class FormioForm extends FormioComponents {
     }
     else {
       return this.executeSubmit();
+    }
+  }
+
+  submitUrl(URL,headers){
+    if(!URL) {
+      return console.warn('Missing URL argument');
+    }
+
+    let submission = this.submission || {};
+    let API_URL  = URL;
+    let settings = {
+      method: 'POST',
+      headers: {}
+    };
+
+    if (headers && headers.length > 0) {
+      headers.map((e) => {
+        if (e.header !== '' && e.value !== '') {
+          settings.headers[e.header] = e.value;
+        }
+      });
+    }
+    if (API_URL && settings) {
+      try {
+        Formio.makeStaticRequest(API_URL,settings.method,submission,settings.headers).then((res) =>{
+          this.emit('requestDone');
+          this.setAlert('success', '<p> Success </p>');
+        })
+      }
+      catch(e) {
+        this.showErrors(e.statusText + ' ' + e.status);
+        this.emit('error',e.statusText + ' ' + e.status);
+        console.error(e.statusText + ' ' + e.status);
+      }
+    }
+    else {
+      this.emit('error', 'You should add a URL to this button.');
+      this.setAlert('warning', 'You should add a URL to this button.');
+      return console.warn('You should add a URL to this button.');
     }
   }
 }