+++ src/utils/utils.spec.js
@@ -2,6 +2,7 @@
 
 /* eslint-env mocha */
 import { expect } from 'chai';
+import _ from 'lodash';
 import writtenNumber from 'written-number';
 import utils from './index';
 import components from './fixtures/components.json';
@@ -258,6 +259,21 @@ describe('checkCondition', () => {
     expect(utils.checkCondition({}, null, {})).to.be.equal(true);
   });
 
+  it('should calculate simple triggers', () => {
+    const component = {
+      key: 'sum',
+      conditional: {
+        when: 'test',
+        eq: 3,
+        show: true
+      }
+    };
+    const data1 = { test: 3 };
+    const data2 = { test: 5 };
+    expect(utils.checkCondition(component, null, data1)).to.be.equal(true);
+    expect(utils.checkCondition(component, null, data2)).to.be.equal(false);
+  });
+
   it('should be able to calculate condition based on javascript code', () => {
     const component = {
       key: 'sum',
@@ -316,5 +332,224 @@ describe('getDateSetting', () => {
 
     expect(utils.getDateSetting(validMomentExpression)).to.be.eql(validDate);
     expect(utils.getDateSetting(invalidMomentExpression)).to.be.equal(null);
+  });
+});
+
+describe('checkTrigger', () => {
+  it('should default to false', () => {
+    expect(utils.checkCondition({}, {type: 'none'}, null, {})).to.be.equal(true);
+  });
+
+  it('should calculate simple triggers', () => {
+    const component = {
+      key: 'sum'
+    };
+    const trigger = {
+      type: 'simple',
+      simple: {
+        when: 'test',
+        eq: 3,
+        show: true
+      }
+    };
+    const data1 = { test: 3 };
+    const data2 = { test: 5 };
+    expect(utils.checkTrigger(component, trigger, null, data1)).to.be.equal(true);
+    expect(utils.checkTrigger(component, trigger, null, data2)).to.be.equal(false);
+  });
+
+  it('should be able to calculate trigger based on javascript code', () => {
+    const component = {
+      key: 'sum'
+    };
+    const trigger = {
+      type: 'javascript',
+      javascript: 'result = data.test === 3'
+    };
+    const data1 = { test: 3 };
+    const data2 = { test: 5 };
+
+    expect(utils.checkTrigger(component, trigger, null, data1)).to.be.equal(true);
+    expect(utils.checkTrigger(component, trigger, null, data2)).to.be.equal(false);
+  });
+
+  it('should be able to calculate trigger based on json logic', () => {
+    const component = {
+      key: 'sum'
+    };
+    const trigger = {
+      type: 'json',
+      json: {
+        '===': [
+          { '_sum': { var: 'data.test' } },
+          6
+        ]
+      }
+    };
+    const data1 = { test: [ 1, 2, 3 ] };
+    const data2 = { test: [ 1, 2, 4 ] };
+
+    expect(utils.checkTrigger(component, trigger, null, data1)).to.be.equal(true);
+    expect(utils.checkTrigger(component, trigger, null, data2)).to.be.equal(false);
+  });
+});
+
+describe('setActionProperty', () => {
+  it('should set a boolean action property to true', () => {
+    const component = {
+      key: 'test',
+      disabled: false
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Disabled',
+        value: 'disabled',
+        type: 'boolean'
+      },
+      state: true
+    };
+    utils.setActionProperty(component, action, {}, {}, true);
+    expect(component.disabled).to.be.equal(true);
+  });
+
+  it('should set a boolean action property to false', () => {
+    const component = {
+      key: 'test',
+      disabled: true
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Disabled',
+        value: 'disabled',
+        type: 'boolean'
+      },
+      state: false
+    };
+    utils.setActionProperty(component, action, {}, {}, true);
+    expect(component.disabled).to.be.equal(false);
+  });
+
+  it('should set a boolean action nested property', () => {
+    const component = {
+      key: 'test',
+      validate: {
+        required: true
+      }
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Required',
+        value: 'validate.required',
+        type: 'boolean'
+      },
+      state: false
+    };
+    utils.setActionProperty(component, action, {}, {}, true);
+    expect(component.validate.required).to.be.equal(false);
+  });
+
+  it('should set a string action property', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Label',
+        value: 'label',
+        type: 'string'
+      },
+      text: 'bar'
+    };
+    utils.setActionProperty(component, action, {}, {}, true);
+    expect(component.label).to.be.equal('bar');
+  });
+
+  it('should set a string action property with row templating', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Label',
+        value: 'label',
+        type: 'string'
+      },
+      text: 'bar {{ row.field }}'
+    };
+    utils.setActionProperty(component, action, {field: 'baz'}, {}, true);
+    expect(component.label).to.be.equal('bar baz');
+  });
+
+  it('should set a string action property with data templating', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Label',
+        value: 'label',
+        type: 'string'
+      },
+      text: 'bar {{ data.field }}'
+    };
+    utils.setActionProperty(component, action, {}, {field: 'baz'}, true);
+    expect(component.label).to.be.equal('bar baz');
+  });
+
+  it('should set a string action property with result templating', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Label',
+        value: 'label',
+        type: 'string'
+      },
+      text: 'bar {{ result }}'
+    };
+    utils.setActionProperty(component, action, {}, {}, 'baz');
+    expect(component.label).to.be.equal('bar baz');
+  });
+
+  it('should set a string action property with component templating', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const action = {
+      type: 'property',
+      property: {
+        label: 'Label',
+        value: 'label',
+        type: 'string'
+      },
+      text: 'bar {{ component.key }}'
+    };
+    utils.setActionProperty(component, action, {}, {}, 'baz');
+    expect(component.label).to.be.equal('bar test');
+  });
+
+  it('should do nothing with a bad request', () => {
+    const component = {
+      key: 'test',
+      label: 'foo'
+    };
+    const originalComponent = _.cloneDeep(component);
+    const action = {
+      type: 'foo',
+    };
+    expect(component).to.deep.equal(originalComponent);
   });
 });