+++ Changelog.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.28.6
+### Added 
+ - Autofocus capability.
+ - Ability to provide spellcheck parameter to input.
+
+### Fixed
+ - Some issues with subforms when performing calcuated values
+ - Problems with subforms performing the load when they are not conditionally available.
+
+## 2.28.5
+### Fixed
+ - Issue with input mask crashing when no input mask is on the field.
+
+## 2.28.4
+### Added
+ - Mask validator for Phone Number.
+
+### Fixed
+ - Default value for component with input mask.
+
+## 2.28.3
+### Changed
+ - The conditional logic where a parent that is conditionally invisible is overridden by a child conditionally visible.
+   This logic is different from the Angular 1 renderer, so we made it consistent where a conditionally visible child will not
+   override a conditionally hidden parent. However, this behavior can be changed by providing the "conditional.overrideParent" flag
+   on the child component.
+
+## 2.28.2
+### Fixed
+ - Issues with the sub-form component not loading the proper source for remote servers.
+ - Issues with the sub-form component not passing along sub data to conditional checks properly.
+
 ## 2.28.1
 ### Fixed
  - Text mask dependencies
@@ -37,7 +69,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - Support for OAuth buttons in the renderer.
  - Ability to add the "Add Another" button on datagrid to either above or below the grid.
- 
+
 ### Fixed
  - Problem where a padding-right is applied to all has-feedback inputs even though an icon is not used.